8/12 Morning Session
Tested thoroughly the correctness of the following actions: OpenANewCourse, AddStudent and ParticipatingInCourse, made various small adjustments as was needed. Added the method GetRawActors() to ActorThreadPool.java, as it was needed in order to add correctly new actors from the implemented actions. And finally, added a mechanism to make certain that Phase 1 is completed before the beginning of phases 2 and 3.